move content to right folders make components SSR friendly
