--- components/appDefinitionContent/index.vue
@@ -1,15 +1,0 @@
-<template>
-  <dd class="app-definition-content">
-    <slot />
-  </dd>
-</template>
-
-<script>
-export default {
-  name: 'AppDefinitionContent'
-}
-</script>
-
-<style scoped>
-
-</style>
