--- components/appDefinitionTitle/index.vue
@@ -1,15 +1,0 @@
-<template>
-  <dt class="app-definition-title">
-    <slot />
-  </dt>
-</template>
-
-<script>
-export default {
-  name: 'AppDefinitionTitle'
-}
-</script>
-
-<style scoped lang="scss">
-
-</style>
