--- components/appInput/index.vue
@@ -1,13 +1,0 @@
-<template>
-  <div class="app-input">
-    <slot />
-  </div>
-</template>
-<script>
-export default {
-  name: 'AppInput'
-}
-</script>
-<style>
-.app-input {}
-</style>
