--- components/appInputSearch/index.vue
@@ -1,13 +1,0 @@
-<template>
-  <div class="app-input-search">
-    <slot />
-  </div>
-</template>
-<script>
-export default {
-  name: 'AppInputSearch'
-}
-</script>
-<style>
-.app-input-search {}
-</style>
