--- components/appSearch/index.vue
@@ -1,13 +1,0 @@
-<template>
-  <div class="app-search">
-    <slot />
-  </div>
-</template>
-<script>
-export default {
-  name: 'AppSearch'
-}
-</script>
-<style>
-.app-search {}
-</style>
