+++ components/networkName/index.vue
@@ -25,10 +25,8 @@ export default {}
       @extend %face-uppercase-xs;
       font-weight:500;
       padding:1rem 1.2rem;
-      margin:1.2rem 1.2rem 0 0;
       background-color: $color-primary;
       border-radius: 0.3rem;
     }
-
   }
 </style>
