--- layouts/index.vue
@@ -1,38 +1,0 @@
-<!--
-<template>
-  <div
-    id="app"
-    class="app"
-  >
-    <AppHeader />
-    <AppView>
-      <RouterView />
-    </AppView>
-  </div>
-</template>
-<script>
-
-import AppHeader from '~/partials/appHeader'
-import AppView from '~/partials/appView'
-
-export default {
-  layout: 'default',
-  name: 'App',
-  components: {
-    AppHeader,
-    AppView
-  }
-}
-</script>
-<style >
-.app {
-  background-color: #F7FAFC;
-  padding-bottom: 2rem;
-}
-  h1 {
-    font-size: 1.2rem;
-    margin-top: 0;
-    font-weight: normal;
-  }
-</style>
--->
