+++ nuxt.config.js
@@ -27,13 +27,22 @@ module.exports = {
   ** Global CSS
   */
   css: [
+    { src: '@aeternity/aepp-components-3/dist/aepp.components.css' },
+    { src: '@aeternity/aepp-components-3/dist/aepp.fonts.css' },
+    { src: 'styles/index.scss', lang: 'scss' }
   ],
 
   /*
   ** Plugins to load before mounting the App
   */
   plugins: [],
-
+  /*
+    ** Router config
+    */
+  router: {
+    linkActiveClass: 'active-link',
+    linkExactActiveClass: 'exact-active-link'
+  },
   /*
   ** Nuxt.js modules
   */
