+++ partials/generation/index.vue
@@ -65,7 +65,7 @@ export default {
 </script>
 
 <style scoped lang="scss">
-  @import "~@aeternity/aepp-components-3/src/styles/variables/colors";
+  @import "../../node_modules/@aeternity/aepp-components-3/src/styles/variables/colors";
   .generation {
     background-color: #FFFFFF;
     display: flex;
