+++ partials/generations/index.vue
