+++ partials/transactions/txListItem/index.vue
@@ -118,7 +118,7 @@ export default {
 </script>
 
 <style scoped lang="scss">
-  @import "../~@aeternity/aepp-components-3/src/styles/variables/colors";
+@import "../../../node_modules/@aeternity/aepp-components-3/src/styles/variables/colors";
   .transaction {
     background-color: #FFFFFF;
     display: flex;
