+++ styles/index.scss
@@ -6,13 +6,13 @@
 /**
  * Global Imports
  */
-@import "../../node_modules/@aeternity/aepp-components-3/src/styles/globals/mixins";
-@import "../../node_modules/@aeternity/aepp-components-3/src/styles/globals/functions";
-@import "../../node_modules/@aeternity/aepp-components-3/src/styles/placeholders/typography";
-@import "../../node_modules/@aeternity/aepp-components-3/src/styles/variables/animations";
-@import "../../node_modules/@aeternity/aepp-components-3/src/styles/variables/colors";
-@import "../../node_modules/@aeternity/aepp-components-3/src/styles/variables/paths";
-@import "../../node_modules/@aeternity/aepp-components-3/src/styles/variables/typography";
+@import "~@aeternity/aepp-components-3/src/styles/globals/mixins";
+@import "~@aeternity/aepp-components-3/src/styles/globals/functions";
+@import "~@aeternity/aepp-components-3/src/styles/placeholders/typography";
+@import "~@aeternity/aepp-components-3/src/styles/variables/animations";
+@import "~@aeternity/aepp-components-3/src/styles/variables/colors";
+@import "~@aeternity/aepp-components-3/src/styles/variables/paths";
+@import "~@aeternity/aepp-components-3/src/styles/variables/typography";
 /**
  * Local Imports
  */
